\docType{class}
\name{DlmSamples}
\alias{DlmSamples}
\title{DLM Samples class}
\usage{
  DlmSamples(...)
}
\description{
  Samples of observations and states generated from a DLM.
  These objects are usually generated by
  \code{\link{dlm_recursion}}.
}
\section{Slots}{
  \describe{ \item{\code{a}}{\code{matrix} containing the
  latent states} \item{\code{y}}{\code{matrix} containing
  the observations} }
}
\seealso{
  \code{\link{dlm_recursion}} which returns object of this
  class.
}

