\name{DLM}
\alias{DLM}
\title{Create DLM objects}
\usage{
  DLM(T, HH, Z, GG, a1 = NULL, P1 = NULL, cc = NULL,
    dd = NULL, HG = NULL)
}
\arguments{
  \item{T}{\code{matrix} of dimension m x m.}

  \item{HH}{\code{matrix} of dimension m x m.}

  \item{Z}{\code{matrix} of dimension N x m.}

  \item{GG}{\code{matrix} of dimension N x N.}

  \item{a1}{\code{numeric} of length m.}

  \item{P1}{\code{matrix} of dimension m x m.}

  \item{cc}{\code{numeric} of dimension m.}

  \item{dd}{\code{numeric} of dimension N.}

  \item{HG}{\code{matrix} of dimension m x N.}
}
\description{
  Create DLM objects
}

