\docType{class}
\name{DlmSamples}
\alias{DlmSamples}
\alias{DlmSamples-class}
\title{DLM Samples class}
\usage{
  DlmSamples(...)
}
\arguments{
  \item{...}{Data to include in the new object.}
}
\description{
  Samples of observations and states generated from a DLM.
  These objects are usually generated by
  \code{\link{dlm_recursion}}.

  The function \code{DlmSamples} is a generator function
  for objects of this class.
}
\section{Slots}{
  \describe{ \item{\code{a}}{\code{Matrix} containing the
  latent states} \item{\code{y}}{\code{Matrix} containing
  the observations} }
}
\seealso{
  \code{\link{dlm_recursion}} which returns object of this
  class.
}

