\docType{class}
\name{MatrixList-class}
\alias{MatrixList-class}
\alias{Ops,Matrix,MatrixList-method}
\alias{Ops,MatrixList,Matrix-method}
\alias{Ops,MatrixList,MatrixList-method}
\alias{cbind2,Matrix,MatrixList-method}
\alias{cbind2,MatrixList,Matrix-method}
\alias{cbind2,MatrixList,MatrixList-method}
\alias{dim,MatrixList-method}
\alias{rbind2,Matrix,MatrixList-method}
\alias{rbind2,MatrixList,Matrix-method}
\alias{rbind2,MatrixList,MatrixList-method}
\title{List of Matrix objects}
\description{
  A list of \code{\linkS4class{Matrix}} objects, which must
  all have the same dimensions.
}
\section{Extends}{
  \describe{ \item{\code{list}}{Directly.} }
}

